Tests Pass
All the first seven tests that I made up pass my initial sanity tests
and I am happy because this is essentially a working version of the
code. I have currently tested it in google-chrome and firefox and both
show the correct results. Infact firefox shows a few warnings for the
javascript as it currently stands.