Added in colour support.
Colour support has now been added in which allows people to colour every
item based on it's value and people can see the evaluation tree take
place as they go. It is now intuitively obvious where the fails in the
systems are. Colors now work perfectly. With gradients and everything.