Extended for Old HTML and Updated tests page.
I have finished the Old HTML code so that it displays details and it
works perfectly on google chrome and old half as well on mozilla
firefox. I have to figure out what the problem is there but essentially
it is done. The Client side of the code base is oficially complete and
seems to work very nicely.

I have alos updated the TODO file and the tests so that you can see both
the HTML5 and Old Html on the one page.